% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{augment}
\alias{augment}
\alias{augment.tbl_esm_eth}
\title{Augment imported file with meta data}
\usage{
augment(x, ...)

\method{augment}{tbl_esm_eth}(x, Parts = NULL, Items = NULL, ...)
}
\arguments{
\item{x}{csv file name}

\item{Parts}{data frame with traits}

\item{Items}{data frame item meta data}
}
\value{
ESM table (tbl_esm)
}
\description{
adds participant and item level meta data
}
\author{
Martin Schmettow
}
