% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{export_item_labels}
\alias{export_item_labels}
\title{Export empty meta data for scales and items}
\usage{
export_item_labels(tbl_esm)
}
\arguments{
\item{tbl_esm}{ESm table produced by read_x_csv}
}
\value{
data frame
}
\description{
Creates an empty table with filled in raw item labels.
}
\author{
Martin Schmettow
}
