% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{export_items_template}
\alias{export_items_template}
\title{Export empty meta data for scales and items}
\usage{
export_items_template(tbl_esm_raw)
}
\arguments{
\item{tbl_esm_raw}{ESm table produced by read_x_csv}
}
\value{
data frame
}
\description{
Creates an empty table with filled in raw item labels.
}
\author{
Martin Schmettow
}
