% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tbl_esm}
\alias{plot.tbl_esm}
\title{Overview plot}
\usage{
\method{plot}{tbl_esm}(x)
}
\arguments{
\item{x}{tbl_esm}
}
\value{
None
}
\description{
returns a long ESM table
}
\author{
Martin Schmettow
}
