% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_ethica_csv}
\alias{read_ethica_csv}
\alias{print.tbl_esm}
\title{Import Ethica CSV file}
\usage{
read_ethica_csv(file)

\method{print}{tbl_esm}(x)
}
\arguments{
\item{file}{csv file name}
}
\value{
ESM table (tbl_esm)
}
\description{
returns a long ESM table
}
\author{
Martin Schmettow
}
