% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{augment}
\alias{augment}
\alias{augment.tbl_esm_eth}
\title{Augment imported file with meta data}
\usage{
augment(x, meta, ...)

\method{augment}{tbl_esm_eth}(x, meta, ...)
}
\arguments{
\item{file}{csv file name}
}
\value{
ESM table (tbl_esm)
}
\description{
returns a long ESM table
}
\author{
Martin Schmettow
}
